==> codebase/04_create_enlistment_records/03_process_year_of_birth_enlistment.Rmd <==
---
title: "3-process-year-of-birth-enlistment"
output: html_document
---


```{r}
library(tidyverse) ## ggplot, dplyr, readr, etc. 
library(dplyr) # %>%, group_by, tally, print
library(readr) # read_csv, write_csv
library(lubridate) #as_date
```

#Set path to temp data & outpath
```{r}
path_to_temp_data <- "/global/scratch/p2p3/pl1_demography/censoc/workspace/enlistment_records/temp2.csv"
out_path          <- "/global/scratch/p2p3/pl1_demography/censoc/workspace/enlistment_records"

```


# Read in temp2
```{r}
data_combined <- read_csv(path_to_temp_data, 
                          col_types = cols(
                            empty_field = col_character(),
                            term_or_enlistment = col_character(),
                            longevity = col_character(),
                            education = col_character(),
                            marital_status = col_character(),
                            component = col_character(),
                            educ_spec = col_character(),
                            defer_date_mmyy = col_character(),
                            mname_clean = col_character(),
                            sex = col_integer(),
                            sex_r = col_integer(),
                            fname_std = col_character()
                            )
                          )

```

# Process year_of_birth

```{r}
min(data_combined$year_of_birth, na.rm=T) # "00"
max(data_combined$year_of_birth, na.rm=T) # "99"

data_combined %>%
  group_by(year_of_birth) %>%
  tally() %>%
  print(n=101)

split_date_of_enlistment_ddmmyy <- function(row) {
  dd <- substr(row, 1, 2)
  mm <- substr(row, 3, 4)
  yy <- substr(row, 5, 6)
  return(list(dd=dd, mm=mm, yy=yy))
}
```

#Process Date of Enlistment DD MM YY
```{r}
res <- lapply(data_combined$date_of_enlistment_ddmmyy, split_date_of_enlistment_ddmmyy) # SLOW
data_combined$date_of_enlistment_dd <- unlist(res)[attr(unlist(res), "names") == "dd"]
data_combined$date_of_enlistment_mm <- unlist(res)[attr(unlist(res), "names") == "mm"]
data_combined$date_of_enlistment_yy <- unlist(res)[attr(unlist(res), "names") == "yy"]
rm(res)

min(data_combined$date_of_enlistment_yy, na.rm=T) # "00"
max(data_combined$date_of_enlistment_yy, na.rm=T) # "99"

data_combined %>%
  group_by(date_of_enlistment_yy) %>%
  tally() %>%
  print(n=100)

nrow(subset(data_combined, date_of_enlistment_yy < "38" | date_of_enlistment_yy > "46")) # 41896

determine_yob <- function(row) {
  yob <- row["year_of_birth"]
  doe <- row["date_of_enlistment_yy"]
  if (!is.na(yob) & !is.na(doe)) {
    yob18 <- as.integer(paste0("18", yob))
    yob19 <- as.integer(paste0("19", yob))
    doe18 <- as.integer(paste0("18", doe))
    doe19 <- as.integer(paste0("19", doe))
    diff1918 <- doe19 - yob18
    diff1919 <- doe19 - yob19
    possible <- c()
    if (diff1918 > 16 & diff1918 < 65) {possible <- c(possible, yob18)}
    if (diff1919 > 16 & diff1919 < 65) {possible <- c(possible, yob19)}
    if (length(possible) != 0) {return(max(possible))}
    else {return(NA)}
  } else {
    return(NA)
  }
}

data_combined$yob_precise <- apply(data_combined, 1, determine_yob) # SLOW

sum(is.na(data_combined$yob_precise)) # 98742
```

```{r}
## create new date variable 
data_combined <- data_combined %>% 
  mutate(date_of_enlistment_yy = as.numeric(date_of_enlistment_yy) + 1900) %>% 
  mutate(date_enlistment_recode = make_date(date_of_enlistment_yy, date_of_enlistment_mm, date_of_enlistment_dd))

## check dates look reasonable and are in yyyy-mm-dd format (match v1 release)
data_combined %>%
  dplyr::count(date_enlistment_recode) %>% 
  arrange(desc(n))

## visualize distribution (loooks reasonable - identical pattern to v1 release)
enliste <- data_combined %>%
  filter(date_of_enlistment_yy %in% c(1938:1947)) %>% ## restrict to years around WWII 
  mutate(date_short = format(as.Date(date_enlistment_recode), "%Y-%m")) %>% 
    group_by(date_short) %>% 
    tally() %>% 
    filter(!is.na(date_short)) %>%
  ggplot(aes(x=date_short, y=n, group=1)) +
  geom_line() +
  geom_point() +
  scale_x_discrete(breaks = paste0(seq(1938, 1947, 1), "-01")) +
  theme_minimal(base_size = 15) +
  ggtitle("Enlistment Date (Month and Year Only)") +
  labs(x = "Date (yyyy-mm)", 
       y = "Count") + 
  scale_y_continuous(labels = scales::comma) +
  theme(axis.text.x = element_text(angle = 45))
```
#save image for tech report
```{r}
library(here)
ggsave(plot = enliste, filename = here("codebase/04_create_enlistment_records/dateofenlist.png"), height = 5, width = 9)
```

```{r}
# Save processed dataset
write_csv(data_combined, paste0(out_path, "/temp3.csv"))
rm(list = ls())

```
