==> codebase/04_create_enlistment_records/09_final_clean_enlistment.Rmd <==
---
title: "9-final-clean-enlistment"
output: html_document
date: "2023-03-02"
---

```{r}
knitr::opts_chunk$set(echo = TRUE)
```

```{r}
library(dplyr)   # %>%, group_by, tally, print
library(ipumsr)  # read_ipums_ddi, ipums_val_labels
library(readr)   # read_csv, write_csv
library(stringr) # str_trim, str_split
```

#update path
```{r}
path_to_temp_data   <-  "/global/scratch/p2p3/pl1_demography/censoc/workspace/enlistment_records/temp8.csv"
out_path <- "/global/scratch/p2p3/pl1_demography/censoc/workspace/enlistment_records/"
```

```{r}
data_combined <- read_csv(path_to_temp_data, 
                          col_types = cols(
                            empty_field = col_character(),
                            term_or_enlistment = col_character(),
                            longevity = col_character(),
                            education = col_character(),
                            marital_status = col_character(),
                            component = col_character(),
                            educ_spec = col_character(),
                            defer_date_mmyy = col_character(),
                            mname_clean = col_character(),
                            sex = col_integer(),
                            sex_r = col_integer(),
                            fname_std = col_character(),
                            `res_state_recode$residence_state_fips` = col_character()
                            )
                          )
```

#Create final df using the subset function and rename columns
```{r}
colnames(data_combined)
```

```{r}
enlistment_records_final_v1 <- subset(data_combined, select = c( "serial_number",  "yob_precise", "fname_clean", "fname_clean_std" , "mname_clean" ,"lname_clean", "sex_r", "date_enlistment_recode", "nativity_recode$nativity_recode" ,  "res_state_recode$state_recode", "res_state_recode$residence_state_fips", "residence_county", "place_of_enlistment","education_recode$education_recode","grade_recode$grade_code_recode",  "branch_recode$branch_code_recode", "term_or_enlistment_recode$term_or_enlistment_recode", "race_recode$race_recode", "citizenship_recode$citizenship_recode", "civ_occupation_recode" ,"marital_status_recode$marital_status_recode", "height_recode", "weight_before_1943","weight_or_AGCT", "component_recode$component_recode","source_recode$source_recode", "component_recode$enlistment_status"))

```

#test df by looking at the first 200 rows
```{r}
head(enlistment_records_final_v1, 200)
```
#rename columns
```{r}
enlistment_records_final <- enlistment_records_final_v1 %>% 
        rename(  
               "fname" = "fname_clean",
               "mname" = "mname_clean",
               "lname" = "lname_clean",
               "sex" = "sex_r",
               "residence_state"="res_state_recode$state_recode",
               "residence_state_fips"="res_state_recode$residence_state_fips",
               "residence_county_fips"="residence_county",
               "byear" = "yob_precise",
               "bpl"="nativity_recode$nativity_recode"  ,
               "date_of_enlistment" = "date_enlistment_recode",
               "source" = "source_recode$source_recode" ,
               "grade_code" = "grade_recode$grade_code_recode" , 
               "term_of_enlistment" = "term_or_enlistment_recode$term_or_enlistment_recode" , 
               "branch_code" = "branch_recode$branch_code_recode" ,
               "component" = "component_recode$component_recode" , 
               "race" = "race_recode$race_recode", 
               "citizenship" = "citizenship_recode$citizenship_recode", 
               "education" = "education_recode$education_recode" ,
               "marital_status" = "marital_status_recode$marital_status_recode" , 
               "height" = "height_recode",
               "weight_before_march_1943" = "weight_before_1943",
               "civilian_occupation" = "civ_occupation_recode",
               "enlistment_status"  = "component_recode$enlistment_status")
```

```{r}
head(enlistment_records_final, 200)
```



# remove rows that contain no actual data 
About 160,000 rows contain the string "ONE OR MORE CARDS WERE UNREADABLE AT THIS LOCATION",
indicating a gap in the data rather than an actual data record.
9,200,232 rows --> 9,039,840 rows
```{r}
enlistment_records_final <- enlistment_records_final %>%
  filter(serial_number != "ONE OR M")
```



# Filter Dates (enlistment & birth)
Birth dates: no one could legally enlist under age 17. We filter to those born in 1930 or earlier, who would have turned 17 in the year 1947. No clear maximum age, but we will filter out birth years before 1860 (cohort that turned 78 in 1938). Records before this date mostly have invalid/missing names and dates of enlistment, so are likely not reliable.

There may cases where (year of enlistment) - (year of birth) < 17, we have not filtered these out.
```{r}
mean(enlistment_records_final$byear %in% 1860:1930) # 98.9% of people are born 1860 - 1930
enlistment_records_final <- enlistment_records_final %>% 
  mutate(byear = ifelse(byear %in% 1860:1930, byear, NA_integer_))
table(enlistment_records_final$byear, useNA = "always")
```

Date of enlistment:
remove anything outside 1938-1947 range
```{r}
table(lubridate::year(enlistment_records_final$date_of_enlistment), useNA = "always")
enlistment_records_final <- enlistment_records_final %>% 
  mutate(date_of_enlistment = if_else(date_of_enlistment >= "1938-01-01" & date_of_enlistment <= "1947-12-31",
                                      date_of_enlistment, as.Date(NA)))
table(lubridate::year(enlistment_records_final$date_of_enlistment), useNA = "always")
```

# pad state fips codes with leading zeros
```{r}
enlistment_records_final$residence_state_fips <- str_pad(enlistment_records_final$residence_state_fips,
                                                         2, "left", pad = "0")
```



```{r}
head(enlistment_records_final)
```


#save final dataset
```{r}
# with cleaned names for publication
write_csv(enlistment_records_final %>% select(-c("fname_clean_std")),
          paste0(out_path, "/finalenlistment.csv"))

# with standardized names for matching
write_csv(enlistment_records_final %>% select(-c("fname")),
          paste0(out_path, "/finalenlistment_for_linking.csv"))
```

clean up
```{r}
rm(list = ls())
```
